requirements.txt
Django==1.11.4
djangorestframework==3.8.2
